Add load balancing strategy and dispatch implementation
